fix listing bucket with marker and limit
When listing a bucket with a / delimiter the number of items returned
by LibreS3 and SX are different, this could be fixed by using NextMarker,
however s3cmd doesn't support it.
Without NextMarker the last Key acts as the next marker for continuing listing,
however this means that all prefixes (directories) have to be listed even if it
exceeds MaxKeys.

Turn off the limit when a delimiter is used in listing.